+++ src/main/java/com/yourssu/roomescape/auth/LoginMember.java
@@ -1,4 +1,4 @@
-package com.yourssu.roomescape.member;
+package com.yourssu.roomescape.auth;
 
 public class LoginMember {
     private Long id;
