+++ src/main/java/com/yourssu/roomescape/auth/TokenDto.java
@@ -1,4 +1,4 @@
-package com.yourssu.roomescape.member;
+package com.yourssu.roomescape.auth;
 
 public class TokenDto {
