+++ src/main/java/com/yourssu/roomescape/common/controller/PageController.java
@@ -1,4 +1,4 @@
-package com.yourssu.roomescape;
+package com.yourssu.roomescape.common.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
