+++ src/main/java/com/yourssu/roomescape/member/MemberController.java
@@ -1,6 +1,9 @@
 package com.yourssu.roomescape.member;
 
-import com.yourssu.roomescape.AppConstants;
+import com.yourssu.roomescape.auth.LoginCheckResponse;
+import com.yourssu.roomescape.auth.LoginRequest;
+import com.yourssu.roomescape.auth.TokenDto;
+import com.yourssu.roomescape.common.AppConstants;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
